refactor(types): DRY - event and order types for improved clarity and consistency
