+++ src/lib/types/layout.ts
@@ -1,0 +1,13 @@
+// src/lib/types/layout.ts
+// Shared types for the customer layout components.
+
+export interface UserInfo {
+  id: number;
+  role: 'admin' | 'customer';
+}
+
+export interface CategoryInfo {
+  id: number;
+  name: string;
+  slug: string;
+}
